fixes and changes v.2

+++ src/main/java/com/example/Calculator.java
@@ -14,8 +14,8 @@ public class Calculator {
         return a * b;
     }
 
-    public static double divide(int a, int b) {
-        return (double) a / b;
+    public static float divide(int a, int b) {
+        return a / b;
     }
 
     public static double squaredRoot(int a) {
